+++ src/app/fantasy/nba/playoffs/SeriesPickCard.tsx
@@ -64,6 +64,7 @@ function TeamButton({
     >
       {/* Team logo */}
       {logoUrl && !isTbd ? (
+        // eslint-disable-next-line @next/next/no-img-element
         <img
           src={logoUrl}
           alt=""